Fix XSLT accumulator, variable, and as clusters
Accumulator cluster: sequence-constructor rule bodies, dynamic map/array
invocation, date comparison with untypedAtomic, temp-tree root/path,
initial-template match rules, and end-phase rules. 17/17 runnable passing.

Variable cluster: empty-URI EQName support (Q{}local), initialize globals
before pattern compilation, fn:sum/fn:avg integer/decimal fixes, and
fn:number boolean parsing. 106/108 passing (2 skipped).

as cluster: preserve subtype dynamic type for xs:integer -> xs:decimal,
separate URI promotion for xs:anyURI -> xs:string, and related coercion
cleanup. 99/99 runnable passing (105 skipped).

Full W3C XSLT 3.0 suite: 4025 passed / 1249 failed / 9326 skipped (76.3%).
Unit tests: 877 passed / 0 failed.

+++ tests/Bosak.XPath.Core.Tests/EndToEndTests.cs
@@ -14,6 +14,7 @@
 //                      | Charles Korthout | 0.2   | 19-05-2026     | Added end-to-end tests for string, sequence, and aggregate functions                   |
 //                      | Charles Korthout | 0.3   | 19-05-2026     | Added tests for Intersect, Except, and SimpleMap operators                             |
 //                      | Charles Korthout | 0.4   | 19-05-2026     | Added tests for Map, Array, and Lookup                                                 |
+//                      | Charles Korthout | 0.5   | 13-06-2026     | Updated FunctionCall_Sum to expect xs:integer for all-integer input                    |
 //                      |==================|=======|================|=========================================================================================
 // ===========================================================================================================================================================
 
@@ -623,7 +624,7 @@ public class EndToEndTests
     public void FunctionCall_Sum()
     {
         var result = Evaluate("sum((1, 2, 3))", LoadDocument());
-        Assert.Equal(6m, result.DecimalValue);
+        Assert.Equal(6L, result.IntegerValue);
     }
 
     [Fact]
